requirements.txt

kafka-python
yfinance
pandas
